+++ eslint.config.js
@@ -10,7 +10,8 @@ export default [
   pluginJs.configs.recommended,
   {
     rules: {
-      "no-unused-vars": "warn"
+      "no-unused-vars": "warn",
+      "no-unreachable": "off"
     },
   }
 ];