+++ CHANGELOG.md
@@ -14,6 +14,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add support for cloud LLM providers (Groq/TogetherAI)
 - Add option `-m, --model` to display the model and provider (without any parameters)
 - Add option `-c, --copy` to copy output to clipboard
+- Add `Jina.AI` API key support
+- Add option `-w, --web [search]` to search the web using Tavily API
 
 ## [0.1.3] - 18/03/2025